% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo.R
\name{anls}
\alias{anls}
\title{Creates a non linear squared model of type: y = a * exp(b * x)}
\usage{
anls(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector that should not contain 0 values}
}
\value{
the predicted values of a and b
}
\description{
Takes in x and y vectors and returns the coefficients a and b of the model created
}
\examples{
 x = c(1:14) ; y = c(1880,813,376,161,100,61,31,9,8,2,7,4,3,2)
}
