% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo.R
\name{y_hat}
\alias{y_hat}
\title{Calculates the predicted y values (y_hat) by the model}
\usage{
y_hat(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector that should not contain 0 values}
}
\value{
predicted y values (y_hat) by the model
}
\description{
Takes in x and y vectors and returns the predicted y values (y_hat) by the model
}
\examples{
 x = c(1:14) ; y = c(1880,813,376,161,100,61,31,9,8,2,7,4,3,2)
}
